Export scales to parent iframe
